pyproject.toml
[project]
name = "bonglem-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "pandas>=3.0.1",
]
